* Components/DatabaseTreeView/DatabaseTreeViewController.[h|m]         * NSBundle+Extensions.m         * Article.[h|m]           NS(U)Integer transitions         * GNUMakefile, *plist, bump to 1.1
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2372 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Grr/Components/DatabaseTreeView/DatabaseTreeViewController.h
@@ -69,7 +69,7 @@
  * the specified index for the given item.
  */
 - (id)outlineView: (NSOutlineView *)outlineView
-            child: (int)index
+            child: (NSInteger)index
            ofItem: (id)item;
 
 /**
@@ -83,7 +83,7 @@
  * This is a required method.  Returns the number of children of
  * the given item.
  */
-- (int)outlineView: (NSOutlineView *)outlineView
+- (NSInteger)outlineView: (NSOutlineView *)outlineView
 numberOfChildrenOfItem: (id)item;
 
 /**
